+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add parsing `PivotTableDefinition#data_fields`
 * Add parsing `PivotTableDefinition#page_fields`
+* Add parsing `PivotTableDefinition#row_fields`
 
 ## 0.36.1 (2023-05-22)